requirements.txt
aiohttp==3.6.2
appnope==0.1.0
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
backcall==0.1.0
bleach==3.1.4
cffi==1.14.0
chardet==3.0.4
coordinates==0.4.0
cryptography==2.9.2
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
future==0.18.2
idna==2.9
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
iso8601==0.1.12
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyter-server-proxy==1.3.2
MarkupSafe==1.1.1
mistune==0.8.4
multidict==4.7.5
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.18.3
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.6.1
pynmea2==1.15.0
pyproj==2.6.0
pyrsistent==0.16.0
python-dateutil==2.8.1
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.3
QtPy==1.9.0
Send2Trash==1.5.0
serial==0.0.97
Shapely==1.7.0
simpervisor==0.3
simple-pid==0.2.4
six==1.14.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
txaio==20.4.1
vpython==7.6.1
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
yarl==1.4.2
RPi.GPIO==0.7.0
pymemcache==3.1.1
firebase-admin==4.3.0